refactor: db관련 클래스들의 패키지 이동
