+++ app/src/main/java/woowacourse/movie/database/TicketDao.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.list.adapter
+package woowacourse.movie.database
 
 import androidx.room.Dao
 import androidx.room.Delete
