+++ app/src/main/java/woowacourse/movie/database/TicketDatabase.kt
@@ -1,10 +1,9 @@
-package woowacourse.movie.list.model
+package woowacourse.movie.database
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import woowacourse.movie.list.adapter.TicketDao
 import woowacourse.movie.ticket.model.DbTicket
 
 @Database(entities = [DbTicket::class], version = 1)
